backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0

# MiDaS depth estimation dependencies
torch>=1.9.0
torchvision>=0.10.0
timm>=0.4.12

# 3D reconstruction dependencies
trimesh>=3.15.0
open3d>=0.17.0
